+++ test/knxip_tests/dib_test.py
@@ -9,6 +9,7 @@ from xknx.knxip import (
     DIBSecuredServiceFamilies,
     DIBServiceFamily,
     DIBSuppSVCFamilies,
+    DIBTunnelingInfo,
     DIBTypeCode,
     KNXMedium,
 )
@@ -132,3 +133,28 @@ class TestKNXIPDIB:
         assert dib.supports(DIBServiceFamily.ROUTING)
         assert not dib.supports(DIBServiceFamily.ROUTING, version=2)
         assert dib.supports(DIBServiceFamily.DEVICE_MANAGEMENT)
+
+    def test_dib_tunneling_info(self):
+        """Test parsing of tunneling info."""
+        raw = (
+            b"\x24\x07\x00\xf8\x40\x01\x00\x05\x40\x02\x00\x05\x40\x03\x00\x05"
+            b"\x40\x04\x00\x05\x40\x05\x00\x05\x40\x06\x00\x05\x40\x07\x00\x05"
+            b"\x40\x08\x00\x06"
+        )
+
+        dib = DIB.determine_dib(raw)
+        assert isinstance(dib, DIBTunnelingInfo)
+        assert dib.from_knx(raw) == 36
+
+        assert dib.max_apdu_length == 248
+
+        assert len(dib.slots) == 8
+        for address in ["4.0.1", "4.0.2", "4.0.3", "4.0.4", "4.0.5", "4.0.6", "4.0.7"]:
+            assert dib.slots[IndividualAddress(address)].usable
+            assert not dib.slots[IndividualAddress(address)].authorized
+            assert dib.slots[IndividualAddress(address)].free
+        assert dib.slots[IndividualAddress("4.0.8")].usable
+        assert dib.slots[IndividualAddress("4.0.8")].authorized
+        assert not dib.slots[IndividualAddress("4.0.8")].free
+
+        assert dib.to_knx() == raw